TASK: Refactor to `NodeName::fromString('main')->tetheredNodeAggregateIdByParent($parentId)`
instead of `NodeAggregateId::fromParentNodeAggregateIdAndNodeName`
